Move netpolicy files to a new package
