+++ platform-operator/internal/netpolicy/netpolicy.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2021, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package k8s
+package netpolicy
 
 import (
 	"context"
